move stage drawing to rendering
