+++ lib/layer/source.go
@@ -1,0 +1,6 @@
+package layer
+
+type Source interface {
+	Frames() *FrameForwarder
+	Start() bool
+}
